moved the endpoints to the controllers - removed .route files
