+++ server/src/controllers/users.controller.js
@@ -14,14 +14,18 @@ class UsersController extends CrudRestController {
   registerRoutes() {
     super.registerRoutes();
     this.router.get('/:id/clients', this.getClients.bind(this));
+    this.router.get("/username/:username", this.getByname.bind(this));
+    this.router.get("/email/:email", this.getByEmail.bind(this));
+    this.router.get("/disable/:id", this.disable.bind(this));
+
   }
 
   /**
    * List resources
    */
   list(req, res) {
-    UserModel.find({ active: true }, '-password', (err, user) =>{
-      if(err) {
+    UserModel.find({ active: true }, '-password', (err, user) => {
+      if (err) {
         this._error(res, err.message, 500);
         return
       }
@@ -47,7 +51,7 @@ class UsersController extends CrudRestController {
     }
 
     // check if user exist
-    UserModel.find({ $or : [ { email: req.body.email }, { username: req.body.username } ] }, (err, user) => {
+    UserModel.find({ $or: [{ email: req.body.email }, { username: req.body.username }] }, (err, user) => {
       if (user.length > 0) {
         this._error(res, "User already exist!")
         return
@@ -63,19 +67,19 @@ class UsersController extends CrudRestController {
       });
 
       newUser.save((error, fluffy) => {
-        if(error){
+        if (error) {
           this._error(res, "Failed to create new user.")
         }
 
         const token = PasswordHasher.generateToken(newUser);
         res.status(200).json({
-            status: 200,
-            errorInfo: "",
-            data: {
-                message: "User created!",
-                token: token,
-                user: user
-            }
+          status: 200,
+          errorInfo: "",
+          data: {
+            message: "User created!",
+            token: token,
+            user: user
+          }
         });
       });
     });
@@ -86,7 +90,7 @@ class UsersController extends CrudRestController {
    */
   get(req, res) {
     UserModel.findById(req.params.id, (error, client) => {
-      if (error){
+      if (error) {
         this._error(res, err.message, 500);
         return
       }
@@ -116,7 +120,7 @@ class UsersController extends CrudRestController {
     }
 
     UserModel.findByIdAndUpdate(id, req.body, (error, user) => {
-      if(error){
+      if (error) {
         this._error(res, 'Failed to update user.');
         return;
       }
@@ -124,7 +128,7 @@ class UsersController extends CrudRestController {
       res.status(200).json({
         status: 200,
         data: {
-            message: 'User updated!'
+          message: 'User updated!'
         }
       });
     });
@@ -143,24 +147,110 @@ class UsersController extends CrudRestController {
    * @param {response} res
    */
   getClients(req, res) {
-    if(!req.params.id) {
+    if (!req.params.id) {
       this._error(res, 'id is required', 500);
       return
     }
 
-    UserModel.findById( req.params.id , 'clients', (err, clients) =>{
-        if(err) {
-          this._error(res, err.message);
-          return
-        }
+    UserModel.findById(req.params.id, 'clients', (err, clients) => {
+      if (err) {
+        this._error(res, err.message);
+        return
+      }
 
-        let response = {
-            status: 200,
-            data: clients ? clients.clients : {}
-        }
-        res.status(200).json(response).end();
+      let response = {
+        status: 200,
+        data: clients ? clients.clients : {}
+      }
+      res.status(200).json(response).end();
     }).populate('clients');
   }
-}
 
+  getByEmail(req, res) {
+    var fieldToValidate = ["email"];
+    var errorMessage = ValidationData(fieldToValidate, req.params);
+
+    if (errorMessage != "") {
+      res.status(500).json({
+        status: 500,
+        errorInfo: errorMessage,
+        data: {}
+      }).end();
+      return
+    }
+
+    UserModel.find({ email: req.params.email }, '-password', (err, user) => {
+      let response = {
+        status: 200,
+        errorInfo: "",
+        data: user ? user : {}
+      }
+
+      res.status(200).json(response).end();
+    });
+  }
+
+  getByname(req, res) {
+    var fieldToValidate = ["username"];
+    var errorMessage = ValidationData(fieldToValidate, req.params);
+
+    if (errorMessage !== "") {
+      res.status(500).json({
+        status: 500,
+        errorInfo: errorMessage,
+        data: {}
+      }).end();
+      return;
+    }
+
+    UserModel.findOne({ username: req.params.username }, '-password', (err, user) => {
+      if (err) {
+        res.status(500).json({
+          status: 500,
+          errorInfo: errorMessage,
+          data: {}
+        }).end();
+        return;
+      }
+
+      let response = {
+        status: 200,
+        errorInfo: "",
+        data: user ? user : {}
+      };
+
+      res.status(200).json(response).end();
+    });
+
+  }
+
+  disable(req, res) {
+    var fieldToValidate = ["id"];
+    var errorMessage = ValidationData(fieldToValidate, req.body);
+
+    if (errorMessage != "") {
+      res.status(500).json({
+        status: 500,
+        errorInfo: errorMessage,
+        data: {}
+      }).end();
+      return
+    }
+
+    UserModel.findByIdAndUpdate(req.body.id, { active: false }, (error, user) => {
+      if (error) {
+        this._error(res, "Failed to delete user.")
+        return
+      }
+
+      res.status(200).json({
+        status: 200,
+        errorInfo: "",
+        data: {
+          message: "User disabled!"
+        }
+      })
+    })
+  }
+}
 module.exports = UsersController;
